Update 1559 LLD strategy fees
